move plugins cofigurations to file to avoid circular imports
