+++ packages/plugins/flow/flow/src/index.ts
@@ -1,47 +1,10 @@
 import { Types, PluginFunction } from '@graphql-codegen/plugin-helpers';
 import { parse, printSchema, visit, GraphQLSchema } from 'graphql';
 import { FlowVisitor } from './visitor';
-import { RawTypesConfig } from '@graphql-codegen/visitor-plugin-common';
+import { FlowPluginConfig } from './config';
 
 export * from './visitor';
 export * from './flow-variables-to-object';
-
-export interface FlowPluginConfig extends RawTypesConfig {
-  /**
-   * @name useFlowExactObjects
-   * @type boolean
-   * @description Generates Flow types as Exact types.
-   * @default true
-   *
-   * @example
-   * ```yml
-   * generates:
-   * path/to/file.ts:
-   *  plugins:
-   *    - flow
-   *  config:
-   *    useFlowExactObjects: false
-   * ```
-   */
-  useFlowExactObjects?: boolean;
-  /**
-   * @name useFlowReadOnlyTypes
-   * @type boolean
-   * @description Generates read-only Flow types
-   * @default false
-   *
-   * @example
-   * ```yml
-   * generates:
-   * path/to/file.ts:
-   *  plugins:
-   *    - flow
-   *  config:
-   *    useFlowReadOnlyTypes: true
-   * ```
-   */
-  useFlowReadOnlyTypes?: boolean;
-}
 
 export const plugin: PluginFunction<FlowPluginConfig> = (schema: GraphQLSchema, documents: Types.DocumentFile[], config: FlowPluginConfig) => {
   const header = `/* @flow */\n\n`;
