+++ packages/plugins/java/kotlin/src/index.ts
@@ -1,59 +1,9 @@
 import { parse, GraphQLSchema, printSchema, visit } from 'graphql';
 import { PluginFunction, Types } from '@graphql-codegen/plugin-helpers';
-import { RawConfig, EnumValuesMap } from '@graphql-codegen/visitor-plugin-common';
 import { KotlinResolversVisitor } from './visitor';
 import { buildPackageNameFromPath } from '@graphql-codegen/java-common';
 import { dirname, normalize } from 'path';
-
-export interface KotlinResolversPluginRawConfig extends RawConfig {
-  /**
-   * @name package
-   * @type string
-   * @description Customize the Java package name. The default package name will be generated according to the output file path.
-   *
-   * @example
-   * ```yml
-   * generates:
-   *   src/main/kotlin/my-org/my-app/Resolvers.kt:
-   *     plugins:
-   *       - kotlin
-   *     config:
-   *       package: custom.package.name
-   * ```
-   */
-  package?: string;
-  /**
-   * @name enumValues
-   * @type EnumValuesMap
-   * @description Overrides the default value of enum values declared in your GraphQL schema.
-   *
-   * @example With Custom Values
-   * ```yml
-   *   config:
-   *     enumValues:
-   *       MyEnum:
-   *         A: 'foo'
-   * ```
-   */
-  enumValues?: EnumValuesMap;
-  /**
-   * @name listType
-   * @type string
-   * @default Iterable
-   * @description Allow you to customize the list type
-   *
-   * @example
-   * ```yml
-   * generates:
-   *   src/main/kotlin/my-org/my-app/Types.kt:
-   *     plugins:
-   *       - kotlin
-   *     config:
-   *       listType: Map
-   * ```
-   */
-  listType?: string;
-}
+import { KotlinResolversPluginRawConfig } from './config';
 
 export const plugin: PluginFunction<KotlinResolversPluginRawConfig> = async (schema: GraphQLSchema, documents: Types.DocumentFile[], config: KotlinResolversPluginRawConfig, { outputFile }): Promise<string> => {
   const relevantPath = dirname(normalize(outputFile));
