+++ packages/plugins/typescript/apollo-angular/src/visitor.ts
@@ -1,7 +1,7 @@
 import { ClientSideBaseVisitor, ClientSideBasePluginConfig, DocumentMode, LoadedFragment, indentMultiline } from '@graphql-codegen/visitor-plugin-common';
 import autoBind from 'auto-bind';
 import { OperationDefinitionNode, print, visit, GraphQLSchema, Kind } from 'graphql';
-import { ApolloAngularRawPluginConfig } from './index';
+import { ApolloAngularRawPluginConfig } from './config';
 import { camelCase } from 'camel-case';
 import { Types } from '@graphql-codegen/plugin-helpers';
