+++ packages/plugins/typescript/compatibility/src/selection-set-to-types.ts
@@ -1,7 +1,7 @@
 import { BaseVisitor } from '@graphql-codegen/visitor-plugin-common';
 import { getBaseType } from '@graphql-codegen/plugin-helpers';
 import { SelectionSetNode, isObjectType, isInterfaceType, isNonNullType, isListType, Kind, GraphQLSchema, isUnionType } from 'graphql';
-import { CompatabilityPluginRawConfig } from './index';
+import { CompatabilityPluginRawConfig } from './config';
 import { CompatabilityPluginConfig } from './visitor';
 
 export type SelectionSetToObjectResult = {
