+++ packages/plugins/typescript/graphql-request/src/config.ts
@@ -1,0 +1,4 @@
+import { RawClientSideBasePluginConfig } from '@graphql-codegen/visitor-plugin-common';
+export interface RawGraphQLRequestPluginConfig extends RawClientSideBasePluginConfig {
+  rawRequest?: boolean;
+}
