+++ packages/plugins/typescript/operations/src/config.ts
@@ -1,0 +1,92 @@
+import { RawDocumentsConfig } from '@graphql-codegen/visitor-plugin-common';
+import { AvoidOptionalsConfig } from '@graphql-codegen/typescript';
+
+export interface TypeScriptDocumentsPluginConfig extends RawDocumentsConfig {
+  /**
+   * @name avoidOptionals
+   * @type boolean
+   * @description This will cause the generator to avoid using TypeScript optionals (`?`) on types,
+   * so the following definition: `type A { myField: String }` will output `myField: Maybe<string>`
+   * instead of `myField?: Maybe<string>`.
+   * @default false
+   *
+   * @example Override all definition types
+   * ```yml
+   * generates:
+   * path/to/file.ts:
+   *  plugins:
+   *    - typescript
+   *    - typescript-operations
+   *  config:
+   *    avoidOptionals: true
+   * ```
+   *
+   * @example Override only specific definition types
+   * ```yml
+   * generates:
+   * path/to/file.ts:
+   *  plugins:
+   *    - typescript
+   *  config:
+   *    avoidOptionals:
+   *      inputValue: true
+   *      object: true
+   * ```
+   */
+  avoidOptionals?: boolean | AvoidOptionalsConfig;
+  /**
+   * @name immutableTypes
+   * @type boolean
+   * @description Generates immutable types by adding `readonly` to properties and uses `ReadonlyArray`.
+   * @default false
+   *
+   * @example
+   * ```yml
+   * generates:
+   * path/to/file.ts:
+   *  plugins:
+   *    - typescript
+   *    - typescript-operations
+   *  config:
+   *    immutableTypes: true
+   * ```
+   */
+  immutableTypes?: boolean;
+  /**
+   * @name flattenGeneratedTypes
+   * @type boolean
+   * @description Flatten fragment spread and inline fragments into a simple selection set before generating.
+   * @default false
+   *
+   * @example
+   * ```yml
+   * generates:
+   * path/to/file.ts:
+   *  plugins:
+   *    - typescript
+   *    - typescript-operations
+   *  config:
+   *    flattenGeneratedTypes: true
+   * ```
+   */
+  flattenGeneratedTypes?: boolean;
+  /**
+   * @name noExport
+   * @type boolean
+   * @description Set the to `true` in order to generate output without `export` modifier.
+   * This is useful if you are generating `.d.ts` file and want it to be globally available.
+   * @default false
+   *
+   * @example Disable all export from a file
+   * ```yml
+   * generates:
+   * path/to/file.ts:
+   *  plugins:
+   *    - typescript
+   *  config:
+   *    noExport: true
+   * ```
+   */
+  noExport?: boolean;
+  globalNamespace?: boolean;
+}
