+++ packages/plugins/typescript/urql/src/index.ts
@@ -1,59 +1,9 @@
 import { Types, PluginValidateFn, PluginFunction } from '@graphql-codegen/plugin-helpers';
 import { visit, GraphQLSchema, concatAST, Kind, FragmentDefinitionNode } from 'graphql';
-import { RawClientSideBasePluginConfig, LoadedFragment } from '@graphql-codegen/visitor-plugin-common';
+import { LoadedFragment } from '@graphql-codegen/visitor-plugin-common';
 import { UrqlVisitor } from './visitor';
 import { extname } from 'path';
-
-export interface UrqlRawPluginConfig extends RawClientSideBasePluginConfig {
-  /**
-   * @name withComponent
-   * @type boolean
-   * @description Customized the output by enabling/disabling the generated Component.
-   * @default true
-   *
-   * @example
-   * ```yml
-   * generates:
-   * path/to/file.ts:
-   *  plugins:
-   *    - typescript
-   *    - typescript-operations
-   *    - typescript-urql
-   *  config:
-   *    withComponent: false
-   * ```
-   */
-  withComponent?: boolean;
-  /**
-   * @name withHooks
-   * @type boolean
-   * @description Customized the output by enabling/disabling the generated React Hooks.
-   * @default false
-   *
-   * @example
-   * ```yml
-   * generates:
-   * path/to/file.ts:
-   *  plugins:
-   *    - typescript
-   *    - typescript-operations
-   *    - typescript-urql
-   *  config:
-   *    withHooks: false
-   * ```
-   */
-  withHooks?: boolean;
-
-  /**
-   * @name urqlImportFrom
-   * @type string
-   * @description You can specify module that exports components `Query`, `Mutation`, `Subscription` and HOCs
-   * This is useful for further abstraction of some common tasks (eg. error handling).
-   * Filepath relative to generated file can be also specified.
-   * @default urql
-   */
-  urqlImportFrom?: string;
-}
+import { UrqlRawPluginConfig } from './config';
 
 export const plugin: PluginFunction<UrqlRawPluginConfig> = (schema: GraphQLSchema, documents: Types.DocumentFile[], config: UrqlRawPluginConfig) => {
   const allAst = concatAST(documents.map(v => v.document));
