+++ packages/plugins/typescript/vue-apollo/src/index.ts
@@ -1,58 +1,9 @@
 import { Types, PluginValidateFn, PluginFunction } from '@graphql-codegen/plugin-helpers';
 import { visit, GraphQLSchema, concatAST, Kind, FragmentDefinitionNode } from 'graphql';
-import { RawClientSideBasePluginConfig, LoadedFragment } from '@graphql-codegen/visitor-plugin-common';
+import { LoadedFragment } from '@graphql-codegen/visitor-plugin-common';
 import { VueApolloVisitor } from './visitor';
 import { extname } from 'path';
-
-export interface VueApolloRawPluginConfig extends RawClientSideBasePluginConfig {
-  /**
-   * @name withCompositionFunctions
-   * @type boolean
-   * @description Customized the output by enabling/disabling the generated Vue composition functions.
-   * @default true
-   *
-   * @example
-   * ```yml
-   * generates:
-   * path/to/file.ts:
-   *  plugins:
-   *    - typescript
-   *    - typescript-operations
-   *    - typescript-vue-apollo
-   *  config:
-   *    withCompositionFunctions: true
-   * ```
-   */
-  withCompositionFunctions?: boolean;
-
-  /**
-   * @name vueApolloComposableImportFrom
-   * @type string
-   * @default @vue/apollo-composable
-   */
-  vueApolloComposableImportFrom?: string;
-
-  /**
-   * @name addDocBlocks
-   * @type boolean
-   * @description Allows you to enable/disable the generation of docblocks in generated code.
-   * Some IDE's (like VSCode) add extra inline information with docblocks, you can disable this feature if your prefered IDE does not.
-   * @default true
-   *
-   * @example
-   * ```yml
-   * generates:
-   * path/to/file.ts:
-   *  plugins:
-   *    - typescript
-   *    - typescript-operations
-   *    - typescript-vue-apollo
-   *  config:
-   *    addDocBlocks: true
-   *
-   */
-  addDocBlocks?: boolean;
-}
+import { VueApolloRawPluginConfig } from './config';
 
 export const plugin: PluginFunction<VueApolloRawPluginConfig> = (schema: GraphQLSchema, documents: Types.DocumentFile[], config: VueApolloRawPluginConfig) => {
   const allAst = concatAST(documents.map(s => s.document));
