+++ website/generate-config.js
@@ -12,23 +12,23 @@ const mapping = {
   '../packages/plugins/other/schema-ast/src/index.ts': BASE_DIR + '/schema-ast.md',
   '../packages/plugins/other/time/src/index.ts': BASE_DIR + '/time.md',
   '../packages/plugins/typescript/typescript/src/config.ts': BASE_DIR + '/typescript.md',
-  '../packages/plugins/typescript/compatibility/src/index.ts': BASE_DIR + '/typescript-compatibility.md',
-  '../packages/plugins/typescript/operations/src/index.ts': BASE_DIR + '/typescript-operations.md',
-  '../packages/plugins/typescript/mongodb/src/index.ts': BASE_DIR + '/typescript-mongodb.md',
-  '../packages/plugins/typescript/resolvers/src/index.ts': BASE_DIR + '/typescript-resolvers.md',
-  '../packages/plugins/typescript/apollo-angular/src/index.ts': BASE_DIR + '/typescript-apollo-angular.md',
-  '../packages/plugins/typescript/urql/src/index.ts': BASE_DIR + '/typescript-urql.md',
-  '../packages/plugins/typescript/react-apollo/src/index.ts': BASE_DIR + '/typescript-react-apollo.md',
-  '../packages/plugins/typescript/vue-apollo/src/index.ts': BASE_DIR + '/typescript-vue-apollo.md',
-  '../packages/plugins/typescript/stencil-apollo/src/index.ts': BASE_DIR + '/typescript-stencil-apollo.md',
+  '../packages/plugins/typescript/compatibility/src/config.ts': BASE_DIR + '/typescript-compatibility.md',
+  '../packages/plugins/typescript/operations/src/config.ts': BASE_DIR + '/typescript-operations.md',
+  '../packages/plugins/typescript/mongodb/src/config.ts': BASE_DIR + '/typescript-mongodb.md',
+  '../packages/plugins/typescript/resolvers/src/config.ts': BASE_DIR + '/typescript-resolvers.md',
+  '../packages/plugins/typescript/apollo-angular/src/config.ts': BASE_DIR + '/typescript-apollo-angular.md',
+  '../packages/plugins/typescript/urql/src/config.ts': BASE_DIR + '/typescript-urql.md',
+  '../packages/plugins/typescript/react-apollo/src/config.ts': BASE_DIR + '/typescript-react-apollo.md',
+  '../packages/plugins/typescript/vue-apollo/src/config.ts': BASE_DIR + '/typescript-vue-apollo.md',
+  '../packages/plugins/typescript/stencil-apollo/src/config.ts': BASE_DIR + '/typescript-stencil-apollo.md',
   '../packages/plugins/typescript/document-nodes/src/index.ts': BASE_DIR + '/typescript-document-nodes.md',
   '../packages/plugins/java/apollo-android/src/plugin.ts': BASE_DIR + '/java-apollo-android.md',
-  '../packages/plugins/java/resolvers/src/index.ts': BASE_DIR + '/java-resolvers.md',
-  '../packages/plugins/java/java/src/index.ts': BASE_DIR + '/java.md',
-  '../packages/plugins/java/kotlin/src/index.ts': BASE_DIR + '/kotlin.md',
-  '../packages/plugins/flow/flow/src/index.ts': BASE_DIR + '/flow.md',
+  '../packages/plugins/java/resolvers/src/config.ts': BASE_DIR + '/java-resolvers.md',
+  '../packages/plugins/java/java/src/config.ts': BASE_DIR + '/java.md',
+  '../packages/plugins/java/kotlin/src/config.ts': BASE_DIR + '/kotlin.md',
+  '../packages/plugins/flow/flow/src/config.ts': BASE_DIR + '/flow.md',
   '../packages/plugins/flow/resolvers/src/index.ts': BASE_DIR + '/flow-resolvers.md',
-  '../packages/plugins/flow/operations/src/index.ts': BASE_DIR + '/flow-operations.md',
+  '../packages/plugins/flow/operations/src/config.ts': BASE_DIR + '/flow-operations.md',
   '../packages/plugins/other/introspection/src/index.ts': BASE_DIR + '/introspection.md',
   '../packages/plugins/other/fragment-matcher/src/index.ts': BASE_DIR + '/fragment-matcher.md',
   '../packages/plugins/other/visitor-plugin-common/src/base-visitor.ts': BASE_DIR + '/base-visitor.md',
